+++ tabular/src/autogluon/tabular/models/rf/rf_model.py
@@ -16,114 +16,10 @@ from autogluon.core.utils.utils import normalize_pred_probas
 from autogluon.core.models import AbstractModel
 from autogluon.features.generators import LabelEncoderFeatureGenerator
 
+from .compilers.native import RFNativeCompiler
+from .compilers.onnx import RFOnnxCompiler
+
 logger = logging.getLogger(__name__)
-
-
-class RFOnnxPredictor:
-    def __init__(self, obj, model):
-        self.num_classes = obj.num_classes
-        self.model = model
-
-    def predict(self, X):
-        input_name = self.model.get_inputs()[0].name
-        label_name = self.model.get_outputs()[0].name
-        return self.model.run([label_name], {input_name: X})[0].squeeze()
-
-    def predict_proba(self, X):
-        input_name = self.model.get_inputs()[0].name
-        label_name = self.model.get_outputs()[1].name
-        pred_proba = self.model.run([label_name], {input_name: X})[0]
-        pred_proba = np.array([[r[i] for i in range(self.num_classes)] for r in pred_proba])
-        return pred_proba
-
-
-class RFNativeCompiler:
-    name = 'native'
-    save_in_pkl = True
-
-    @staticmethod
-    def can_compile():
-        return True
-
-    @staticmethod
-    def compile(obj, path: str):
-        from sklearn.ensemble import RandomForestClassifier, RandomForestRegressor
-        from sklearn.ensemble import ExtraTreesRegressor, ExtraTreesClassifier
-        os.makedirs(os.path.dirname(path), exist_ok=True)
-        if isinstance(obj.model, (RandomForestClassifier, RandomForestRegressor,
-                                  ExtraTreesClassifier, ExtraTreesRegressor)):
-            with open(path + 'model_native.pkl', 'wb') as fp:
-                fp.write(pickle.dumps(obj.model))
-            return obj.model
-        else:
-            return RFNativeCompiler.load(obj=obj, path=path)
-
-    @staticmethod
-    def load(obj, path: str):
-        pkl = None
-        with open(path + 'model_native.pkl', 'rb') as fp:
-            pkl = fp.read()
-        return pickle.loads(pkl)
-
-
-class InferenceSessionWrapper:
-    """
-    Wrap around InferenceSession in onnxruntime, since it cannot be pickled.
-    See https://github.com/microsoft/onnxruntime/issues/10097
-    """
-    def __init__(self, onnx_bytes):
-        import onnxruntime as rt
-        self.sess = rt.InferenceSession(onnx_bytes.SerializeToString(),
-                                        providers=['CPUExecutionProvider'])
-
-    def run(self, *args):
-        return self.sess.run(*args)
-
-    def get_inputs(self, *args):
-        return self.sess.get_inputs(*args)
-
-    def get_outputs(self, *args):
-        return self.sess.get_outputs(*args)
-
-    def __getstate__(self):
-        # No need to duplicate the model parameters here.
-        return {}
-
-    def __setstate__(self, values):
-        pass
-
-class RFOnnxCompiler:
-    name = 'onnx'
-    save_in_pkl = False
-
-    @staticmethod
-    def can_compile():
-        try:
-            import skl2onnx
-            return True
-        except ImportError:
-            return False
-
-    @staticmethod
-    def compile(obj, path: str):
-        if isinstance(obj.model, RFOnnxPredictor):
-            return obj.model
-        # Convert into ONNX format
-        from skl2onnx import convert_sklearn
-        from skl2onnx.common.data_types import FloatTensorType
-        initial_type = [('float_input', FloatTensorType([None, obj._num_features_post_process]))]
-        onx = convert_sklearn(obj.model, initial_types=initial_type)
-        os.makedirs(os.path.dirname(path), exist_ok=True)
-        with open(path + "model.onnx", "wb") as f:
-            f.write(onx.SerializeToString())
-        return RFOnnxCompiler.load(obj=obj, path=path)
-
-    @staticmethod
-    def load(obj, path: str):
-        import onnx
-        onnx_bytes = onnx.load(path + "model.onnx")
-        model = InferenceSessionWrapper(onnx_bytes)
-        return RFOnnxPredictor(obj=obj, model=model)
 
 
 class RFModel(AbstractModel):
@@ -334,7 +230,6 @@ class RFModel(AbstractModel):
             model.estimators_ = None
         self.model = model
         self.params_trained['n_estimators'] = self.model.n_estimators
-        self.compile(path=self.path)
 
     # TODO: Remove this after simplifying _predict_proba to reduce code duplication. This is only present for SOFTCLASS support.
     def _predict_proba(self, X, **kwargs):
@@ -473,22 +368,8 @@ class RFModel(AbstractModel):
     def _valid_compilers(self):
         return [RFNativeCompiler, RFOnnxCompiler]
 
-    def _get_compiler(self):
-        compiler = self.params_aux.get('compiler', None)
-        compilers = self._valid_compilers()
-        compiler_names = {c.name: c for c in compilers}
-        if compiler is not None and compiler not in compiler_names:
-            raise AssertionError(f'Unknown compiler: {compiler}. Valid compilers: {compiler_names}')
-        if compiler is None:
-            return RFNativeCompiler
-        compiler_cls = compiler_names[compiler]
-        compiler_fallback = self.params_aux.get('compiler_fallback_to_native', True)
-        if not compiler_cls.can_compile():
-            if not compiler_fallback:
-                raise AssertionError(f'Specified compiler ({compiler}) is unable to '
-                                     'compile (potentially lacking dependencies) and "compiler_fallback==False"')
-            compiler_cls = RFNativeCompiler
-        return compiler_cls
+    def _default_compiler(self):
+        return RFNativeCompiler
 
     def save(self, path: str = None, verbose=True) -> str:
         _model = self.model
